pyproject.toml
[project]
name = "dagster_plus"
version = "0.1.0"
description = "Simple dagster_plus"
authors = [
    { name = "quan.dao", email = "quan.dao@definer-inc.com" },
]
requires-python = ">=3.10,<3.14"
dependencies = [
    "dagster>=1.12.3,<2.0.0",
    "dagster-cloud>=1.12.3,<2.0.0",
    "pandas>=2.3.3,<3.0.0",
    "snowflake-connector-python[pandas]<4.0.0",
    "snowflake-sqlalchemy>=1.7.7,<2.0.0",
    "aiohttp>=3.8.0,<4.0.0",
    "tenacity>=8.2.3,<9.0.0",
]

[project.optional-dependencies]
dev = [
    "dagster-dg-cli",
    "dagster-webserver",
    "pytest",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "dagster_plus"
registry_modules = [
    "dagster_plus.assets.*",
    "dagster_plus.resources.*",
]
